Fix(OutputView): printWinningStatistics 메소드 분리
printWinningStatistics 메소드를 분리하였습니다.

각 등수의 개수를 출력하는 printRanking
수익률을 출력하는 printEarningRate
로 분리하였습니다.

Relate to: #5

+++ src/main/java/lotto/Domain/Rank.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Map;
 
 public class Rank {
-    private int earningRate;
+    private float earningRate;
     private Map<RankType, Integer> rank = new EnumMap<RankType, Integer>(RankType.class);
 
     public void setEarningRate(int money) {
@@ -17,9 +17,9 @@ public class Rank {
                 total += rankType.getPrizeMoney();
             }
         }
-        this.earningRate = total / money * 100;
+        this.earningRate = total / (float)money * 100;
     }
-    public int getEarningRate() {
+    public float getEarningRate() {
         return this.earningRate;
     }
